feat: add search functionality for posts with pagination and sorting, including DTOs for request and response

+++ src/post/post.service.ts
@@ -1,11 +1,16 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Post } from 'src/entities/post.entity';
-import { Repository } from 'typeorm';
+import { Like, Repository } from 'typeorm';
 import { CreatePostDto } from './dto/create-post.dto';
 import { JwtPayload } from 'src/interfaces/jwt-payload.interface';
 import { CategoryService } from 'src/category/category.service';
 import { UpdatePostDto } from './dto/update-post.dto';
+import { SearchPostDto } from './dto/search-post.dto';
 
 @Injectable()
 export class PostService {
@@ -14,6 +19,28 @@ export class PostService {
     private readonly postRepository: Repository<Post>,
     private readonly categoryService: CategoryService,
   ) {}
+
+  async searchPosts(query: SearchPostDto) {
+    const { search, limit, page, sort, order } = query;
+    const where = search ? { title: Like(`%${search}%`) } : {};
+
+    const [data, totalRecord] = await this.postRepository.findAndCount({
+      where,
+      order: { [sort]: order },
+      take: limit,
+      skip: (page - 1) * limit,
+      relations: ['category'],
+    });
+
+    return {
+      data,
+      meta: {
+        totalRecord,
+        currentPage: page,
+        totalPage: Math.ceil(totalRecord / limit),
+      },
+    };
+  }
 
   async findAll(): Promise<Post[]> {
     return this.postRepository.find({
@@ -28,7 +55,6 @@ export class PostService {
       relations: ['category'],
     });
 
-    console.log(post);
     if (!post) {
       throw new NotFoundException(`Post with ID ${id} not found`);
     }
@@ -45,24 +71,29 @@ export class PostService {
     return this.findOne(data.id);
   }
 
-  async update(
-    id: number,
-    post: UpdatePostDto,
-    decode: JwtPayload,
-  ): Promise<Post> {
+  async update(id: number, userId: number, post: UpdatePostDto): Promise<Post> {
+    await this.validatePermission(id, userId);
     if (post?.categoryId) {
       await this.categoryService.findOne(post.categoryId);
     }
+
     await this.postRepository.update(id, {
       ...post,
-      userId: decode.id,
+      userId: userId,
       categoryId: post.categoryId,
     });
     return this.findOne(id);
   }
 
-  async remove(id: number): Promise<void> {
-    await this.findOne(id);
+  async remove(id: number, userId: number): Promise<void> {
+    await this.validatePermission(id, userId);
     await this.postRepository.delete(id);
+  }
+
+  private async validatePermission(id: number, userId: number): Promise<void> {
+    const existingPost = await this.findOne(id);
+    if (existingPost.userId !== userId) {
+      throw new ForbiddenException('Forbidden');
+    }
   }
 }
